detection/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
alabaster=0.7.10=py36hcd07829_0
asn1crypto=0.24.0=py36_0
asteval=0.9.13=pyh24bf2e0_0
babel=2.5.3=py36_0
blas=1.0=mkl
bokeh=0.12.15=py36_0
ca-certificates=2019.3.9=hecc5488_0
certifi=2019.3.9=py36_0
cffi=1.11.5=py36h945400d_0
chardet=3.0.4=py36h420ce6e_1
click=6.7=py36hec8c647_0
cloudpickle=0.5.2=py36_1
cryptography=2.6.1=py36h7a1dbc1_0
cycler=0.10.0=py36h009560c_0
cython=0.28.2=py36hfa6e2cd_0
cytoolz=0.9.0.1=py36hfa6e2cd_0
dask=0.17.2=py36_0
dask-core=0.17.2=py36_0
decorator=4.3.0=py36_0
distributed=1.21.6=py36_0
docutils=0.14=py36h6012d8f_0
freetype=2.8=h51f8f2c_1
grpcio=1.16.1=py36h351948d_1
grpcio-tools=1.16.1=py36ha925a31_0
h5py=2.7.1=py36he54a1c3_0
hdf5=1.10.1=h98b8871_1
heapdict=1.0.0=py36_2
icc_rt=2017.0.4=h97af966_0
icu=58.2=ha66f8fd_1
idna=2.6=py36h148d497_1
imageio=2.3.0=py36_0
imagesize=1.0.0=py36_0
intel-openmp=2018.0.0=8
jinja2=2.10=py36h292fed1_0
jpeg=9b=hb83a4c4_2
kiwisolver=1.0.1=py36h12c3424_0
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.7.1=h7bd577a_0
libtiff=4.0.9=h0f13578_0
llvmlite=0.28.0=py36ha925a31_0
lmfit=0.9.13=pyh24bf2e0_0
locket=0.2.0=py36hfed976d_1
markupsafe=1.0=py36h0e26971_1
matplotlib=2.2.2=py36h153e9ff_1
microscope=0.1.0=pypi_0
mkl=2018.0.2=1
mkl_fft=1.0.1=py36h452e1ab_0
mkl_random=1.0.1=py36h9258bd6_0
msgpack-python=0.5.6=py36he980bc4_0
networkx=2.1=py36_0
numba=0.43.1=py36hf9181ef_0
numpy=1.14.2=py36h5c71026_1
olefile=0.45.1=py36_0
opencv=3.3.1=py36h20b85fd_1
openssl=1.1.1b=hfa6e2cd_2
packaging=17.1=py36_0
pandas=0.22.0=py36h6538335_0
partd=0.3.8=py36hc8e763b_0
pillow=5.1.0=py36h0738816_0
pip=9.0.3=py36_0
protobuf=3.7.1=py36h33f27b4_0
psutil=5.4.5=py36hfa6e2cd_0
pycparser=2.18=py36hd053e01_1
pygments=2.2.0=py36hb010967_0
pyopenssl=17.5.0=py36h5b7d817_0
pyparsing=2.2.0=py36h785a196_1
pyqt=5.9.2=py36h1aa27d4_0
pyro4=4.71=pypi_0
pyserial=3.4=pypi_0
pysocks=1.6.8=py36_0
python=3.6.5=h0c2934d_0
python-dateutil=2.7.2=py36_0
pytoolbox=1.0=pypi_0
pytorch=0.4.0=py36_cuda80_cudnn7he774522_1
pytz=2018.4=py36_0
pywavelets=0.5.2=py36hc649158_0
pyyaml=3.12=py36h1d1928f_1
qt=5.9.7=vc14h73c81de_0
requests=2.18.4=py36h4371aae_1
scikit-image=0.13.1=py36hfa6e2cd_1
scipy=1.0.1=py36hce232c7_0
serpent=1.24=pypi_0
setuptools=39.0.1=py36_0
sip=4.19.8=py36h6538335_0
six=1.11.0=py36h4db2310_1
snowballstemmer=1.2.1=py36h763602f_0
sortedcontainers=1.5.10=py36_0
sphinx=1.7.3=py36_0
sphinxcontrib=1.0=py36hbbac3d2_1
sphinxcontrib-websupport=1.0.1=py36hb5e5916_1
sqlite=3.28.0=he774522_0
tblib=1.3.2=py36h30f5020_0
tifffile=0.14.0=pypi_0
tk=8.6.7=hcb92d03_3
toolz=0.9.0=py36_0
torchvision=0.2.1=pypi_0
tornado=5.0.2=py36_0
typing=3.6.4=py36_0
uncertainties=3.0.3=py36_1000
unknown=0.0.0=pypi_0
urllib3=1.22=py36h276f60a_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.15.26706=h3a45250_4
wheel=0.31.0=py36_0
win_inet_pton=1.0.1=py36he67d7fd_1
wincertstore=0.2=py36h7fe50ca_0
yaml=0.1.7=hc54c509_2
zict=0.1.3=py36h2d8e73e_0
zlib=1.2.11=h8395fce_2
